+++ decoder/packet_dns.h
@@ -275,8 +275,8 @@ list_t *parse_dns_options(struct dns_resource_record *rr);
 void free_dns_options(list_t *opt);
 
 /* internal to the decoder */
-packet_error handle_dns(unsigned char *buffer, int n, struct application_info *info);
-int parse_dns_name(unsigned char *buffer, int n, unsigned char *ptr, char name[]);
+packet_error handle_dns(unsigned char *buffer, size_t n, struct application_info *info);
+int parse_dns_name(unsigned char *buffer, size_t n, unsigned char *ptr, char name[]);
 void free_dns_packet(struct dns_info *dns);
 
 #endif
